Checking the products

+++ repos/fdbt-site/src/pages/api/deleteAllProducts.ts
@@ -7,10 +7,10 @@ import { PRODUCTS_DATA_BUCKET_NAME } from '../../constants';
 
 export default async (req: NextApiRequestWithSession, res: NextApiResponse): Promise<void> => {
     try {
-        if (process.env.NODE_ENV === 'development' || process.env.NODE_ENV === 'test') {
-            const nationalOperatorCode = getAndValidateNoc(req, res);
-            const products = await getAllProductsByNoc(nationalOperatorCode);
+        const nationalOperatorCode = getAndValidateNoc(req, res);
+        const products = await getAllProductsByNoc(nationalOperatorCode);
 
+        if (products.length > 0 && (process.env.NODE_ENV === 'development' || process.env.NODE_ENV === 'test')) {
             const matchingJsonLinks = products.map((product) => product.matchingJsonLink);
             if (matchingJsonLinks.length > 0) {
                 await deleteMultipleObjectsFromS3(matchingJsonLinks, PRODUCTS_DATA_BUCKET_NAME);
